[MOBI-29-faq] Refactor: 상수 파일 케밥 케이스로 변경
