+++ apps/academy/src/widget/landing/constants/faq-list.ts
