+++ apps/academy/src/widget/landing/constants/index.ts
@@ -1,1 +1,2 @@
-export { FAQ_LIST } from './faqList'
+export type { FaqList } from './faq-list'
+export { FAQ_LIST } from './faq-list'
